requirements.txt
discord.py>=2.0.0
python-dotenv
aiofiles
requests
aiosqlite
pytz
aiohttp
Flask>=2.3.2

# Core dependencies
discord.py>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
requests>=2.31.0
aiosqlite>=0.19.0
pytz>=2023.3.post1
aiohttp>=3.8.6
Flask>=2.3.2
Werkzeug>=3.0.1

# Optional but recommended for date parsing and duration handling
python-dateutil>=2.8.2

# Optional scheduler for reminders/daily jobs
APScheduler>=3.10.1

# Testing
pytest>=7.0.0

# Music features
yt-dlp>=2025.10.1
PyNaCl>=1.5.0

# Async features
asyncio>=3.4.3
